Converting Octave Matrix values to JS objects and vice-versa in feval.
Adding eval_string function for use with a CLI to the interpreter.
Adding file system methods to the module so that they do not get minified
by the Closure compiler.
Making the worker build output a single JS file with the WASM and M-files
embedded because VSCode web extensions need to be compiled into a single
file.
Reinstating wasm-extensions in favour of DISABLE_EXCEPTION_CATCHING=0
because VSCode web extensions run in a web-worker environment even with
desktop instances built on Electron and Node.
Adding test page for worker build (requires EXPORT_ES6=0).